Share function definitions for with setup
Because with and setup define many variables, a large test
case could get really slow when compiling, as it would define
many variables in the same module body, causing some chunks
during compilation to become really slow.

To solve this, we store this code in anonymous functions and
reuse it throughout the test case.